Convert to use BigDecimal for prices

+++ CryptoData/app/src/main/java/dev/shorthouse/cryptodata/data/source/remote/model/CoinApiModel.kt
@@ -1,6 +1,7 @@
 package dev.shorthouse.cryptodata.data.source.remote.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class CoinApiModel(
     val id: String,
@@ -8,7 +9,7 @@ data class CoinApiModel(
     val symbol: String,
     val image: String,
     @SerializedName("current_price")
-    val currentPrice: Double,
+    val currentPrice: BigDecimal,
     @SerializedName("price_change_percentage_24h")
     val priceChangePercentage24h: Double,
     @SerializedName("market_cap_rank")
